Move logging setup out of __init__.py
