+++ src/pyprefab/logging.py
@@ -1,0 +1,27 @@
+import sys
+
+import structlog
+
+
+def setup_logging():
+    shared_processors = [
+        structlog.processors.TimeStamper(fmt='%Y-%m-%d %H:%M:%S'),
+        structlog.processors.add_log_level,
+    ]
+
+    if sys.stderr.isatty():
+        # If we're in a terminal, pretty print the logs.
+        processors = shared_processors + [
+            structlog.dev.ConsoleRenderer(),
+        ]  # pragma: no cover
+    else:
+        # Otherwise, output logs in JSON format
+        processors = shared_processors + [
+            structlog.processors.dict_tracebacks,
+            structlog.processors.JSONRenderer(),
+        ]
+
+    structlog.configure(
+        processors=processors,
+        cache_logger_on_first_use=True,
+    )
